Fully Moved to Lumen Added Migrations Updated .env for the new Laravel vagrant box
