+++ .env
@@ -7,8 +7,8 @@ DB_CONNECTION=mysql
 DB_HOST=127.0.0.1
 DB_PORT=3306
 DB_DATABASE=trackmate
-DB_USERNAME=trackmate
-DB_PASSWORD=trackmate
+DB_USERNAME=homestead
+DB_PASSWORD=secret
 
 CACHE_DRIVER=array
 QUEUE_DRIVER=array
