+++ app/Http/Controllers/UserController.php
@@ -7,15 +7,13 @@ namespace App\Http\Controllers;
 
 use Illuminate\Hashing\BcryptHasher;
 use Illuminate\Http\Request;
-use Trackmate\Models\User;
+use App\Models\User;
 
 class UserController extends Controller
 {
     
     public function __construct(BcryptHasher $hasher)
     {
-        $this->middleware('oauth', ['except' => ['index', 'show']]);
-        
         parent::__construct($hasher);
     }
