+++ app/Models/User.php
@@ -1,6 +1,6 @@
 <?php
 
-namespace Trackmate\Models;
+namespace App\Models;
 
 use Illuminate\Auth\Authenticatable;
 use Laravel\Lumen\Auth\Authorizable;
